fix(frontend): resolve ESLint errors in CI pipeline
- Extract useFormField hook to separate file to fix react-refresh/only-export-components
- Replace any types with proper TypeScript interfaces in stack pages
- Add proper type definitions for StackFormData interface
- Ensure form.tsx only exports React components

Fixes 5 ESLint errors:
- react-refresh/only-export-components in form.tsx
- @typescript-eslint/no-explicit-any in create-stack.tsx (2 errors)
- @typescript-eslint/no-explicit-any in edit-stack.tsx (2 errors)
